requirements.txt
pytesseract
Pillow
tkinter
